refactor(arrow): switching the slice utilities to use int instead of int64 (#581)
Using `int` is more common than `int64` for lengths so, even if arrow
uses `int64`, the slice lengths should still use `int`. Additionally,
the arrow `Len()` method for slices uses `int` as the return value so
this is better for that anyway.

+++ arrow/bool.go
@@ -17,8 +17,8 @@ func NewBool(vs []bool, alloc *memory.Allocator) *array.Boolean {
 	return a
 }
 
-func BoolSlice(arr *array.Boolean, i, j int64) *array.Boolean {
-	data := array.NewSliceData(arr.Data(), i, j)
+func BoolSlice(arr *array.Boolean, i, j int) *array.Boolean {
+	data := array.NewSliceData(arr.Data(), int64(i), int64(j))
 	defer data.Release()
 	return array.NewBooleanData(data)
 }
